arm64: dts: qcom: msm8916-samsung-cprime: Add initial device tree
Samsung Galaxy Core Prime is a phone based on MSM8916. It is similar
to the other Samsung devices based on MSM8916 with only a few minor
differences.

The device tree contains initial support for the Core Prime with:
- GPIO keys
- SDHCI (internal and external storage)
- USB Device Mode
- UART
- WCNSS (WiFi/BT)
- Regulators

Core Prime and Grand Prime are similar, with some differences in
accelerometer, MUIC, NFC, panel and touchscreen. The common parts are
shared in msm8916-samsung-gprime-common.dtsi to reduce duplication.

Signed-off-by: Lin, Meng-Bo <linmengbo0689@protonmail.com>