cpufreq: qcom-nvmem: Add MSM8909
The MSM8909 SoC - when used with the PM8909 - needs to scale a single
power domain (here called "apc"). The speedbin that controls the
available frequencies is also easily readable from a nvmem cell without
further processing.

Note that when MSM8909 is paired with other PMICs such as PM8916 or
PM660 there are more power domains that must be scaled correctly.
Handling this is left up for future work.

Signed-off-by: Stephan Gerhold <stephan.gerhold@kernkonzept.com>